Move additional functions to convenience package.
