+++ agrirouter-api-java-api/src/main/java/com/dke/data/agrirouter/api/service/messaging/ListEndpointsService.java
@@ -1,11 +1,5 @@
 package com.dke.data.agrirouter.api.service.messaging;
 
-import com.dke.data.agrirouter.api.dto.onboard.OnboardingResponse;
 import com.dke.data.agrirouter.api.service.parameters.ListEndpointsParameters;
 
-public interface ListEndpointsService extends MessagingService<ListEndpointsParameters> {
-
-  String requestFullListFilteredByAppliedRoutings(OnboardingResponse onboardingResponse);
-
-  String requestFullList(OnboardingResponse onboardingResponse);
-}
+public interface ListEndpointsService extends MessagingService<ListEndpointsParameters> {}
